[Front] Amend directory strcture for src/ directory.
Remove needless level of directory nesting.
