+++ frontend/src/components/CodeEditor/SoloEditor.js
@@ -1,5 +1,5 @@
 import React, {Component} from "react";
-import "../../../CSS/AceEditor.css";
+import "../../CSS/AceEditor.css";
 import code from './SeedCode'
 import brace from 'brace';
 import AceEditor from 'react-ace';
