+++ frontend/src/components/Issues/solvedIssues.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import axios from 'axios';
 import { Link, Route } from 'react-router-dom';
-import '../../.././CSS/Issues.css';
+import '../.././CSS/Issues.css';
 
-class OpenIssues extends React.Component {
+class SolvedIssues extends React.Component {
     constructor() {
         super()
         this.state = {
@@ -16,7 +16,7 @@ class OpenIssues extends React.Component {
             .get('/users/getTicketFeed')
             .then(res => {
                 this.setState({
-                    userdata: res.data.data.filter(issue => issue.problemstatus === '0')
+                    userdata: res.data.data.filter(issue => issue.problemstatus === '1')
                 })
             })
             .catch(err => {
@@ -27,6 +27,7 @@ class OpenIssues extends React.Component {
 
     renderIssues = () => {
         const { userdata } = this.state
+        console.log(`openissue`, userdata)
         return userdata.map((v, i) =>
             <div class="issue" id={`${i}`}>
                 <img src={v.profilepic} />{" "}
@@ -59,4 +60,4 @@ class OpenIssues extends React.Component {
     }
 }
 
-export default OpenIssues;
+export default SolvedIssues;
