+++ frontend/src/components/LandingPage/LaptopCode.js
