+++ frontend/src/components/LoginPage/loginUser.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Link, Route } from 'react-router-dom';
 import axios from 'axios';
-import '../../../CSS/Auth.css';
+import '../../CSS/Auth.css';
 
 
 class LoginUser extends Component {
